doubles

+++ gwv/09/src/PosTag.java
@@ -4,9 +4,9 @@ public class PosTag {
         // Tagname
 	String tagname = null;
         // Übergangswahrscheinlichkeitsmap 
-	HashMap<PosTag, Float> transitions_probs = new HashMap<PosTag, Float>();
+	HashMap<PosTag, Double> transitions_probs = new HashMap<PosTag, Double>();
         // Emissionswahrscheinlichkeitsmap
-	HashMap<String, Float> word_probs = new HashMap<String, Float>();
+	HashMap<String, Double> word_probs = new HashMap<String, Double>();
         // Counter dienen dann als Divisor um die relativen Häufigkeiten zu berechnen
 	int number_words = 0;
 	int number_transitions  = 0;
@@ -20,12 +20,12 @@ public class PosTag {
 		number_transitions++;
 		// System.out.println("In count for tag: "+ this.toString() + " Number_t: " + this.number_transitions);
 		if (transitions_probs.containsKey(tag)){
-			float count =transitions_probs.get(tag);
-			count = count + (float)1.0;
+			Double count =transitions_probs.get(tag);
+			count = count + (Double)1.0;
 			transitions_probs.put(tag,count);
 		}
 		else {
-			transitions_probs.put(tag,(float)(1.0));
+			transitions_probs.put(tag,(Double)(1.0));
 		}
 		
 	}
@@ -33,12 +33,12 @@ public class PosTag {
 	void inc_count_words(String word){
 		number_words++;
 		if (word_probs.containsKey(word)){
-			float count = word_probs.get(word);
-			count = count + (float)1;
+			Double count = word_probs.get(word);
+			count = count + (double)1;
 			word_probs.put(word,count);
 		}
 		else {
-			word_probs.put(word,(float)(1));
+			word_probs.put(word,(double)(1));
 		}
 		
 	}
@@ -49,19 +49,19 @@ public class PosTag {
 	// Normalizes the probabilities for the given tags, i.e. prob of next tag and emission prob for word.  
 	void normalize_probabilities(){
 		// next tag probs
-		for (Entry<PosTag, Float> entry : transitions_probs.entrySet()){
-			float prob  = entry.getValue() / this.number_transitions; 
+		for (Entry<PosTag, Double> entry : transitions_probs.entrySet()){
+			Double prob  = entry.getValue() / this.number_transitions; 
 			transitions_probs.put(entry.getKey(), prob); 
 		}
 		// emissioned word probs
-		for (Entry<String, Float> entry : word_probs.entrySet()){
-			float prob  = entry.getValue() / this.number_words; 
+		for (Entry<String, Double> entry : word_probs.entrySet()){
+			Double prob  = entry.getValue() / this.number_words; 
 			word_probs.put(entry.getKey(), prob); 
 		}
 		// Nur zur Ausgabe der berechneten Wahrscheinlichkeiten und ob diese Sinn machen (Hätte man auch gleich in der obigen Schleife berechnen können) 
-		float test = 0; 
-		for (Entry<String, Float> entry : word_probs.entrySet()){
-			float prob  =entry.getValue(); 
+		double test = 0; 
+		for (Entry<String, Double> entry : word_probs.entrySet()){
+			double prob  = entry.getValue(); 
 			word_probs.put(entry.getKey(), prob);
 			test = test + word_probs.get(entry.getKey()); 
 		}
